requirements.txt
langchain==0.3.24
langchain-community==0.3.23
langchain-core==0.3.56
langchain-ollama==0.3.2
ollama==0.4.8
pypdf==5.4.0
chromadb==0.4.24
beautifulsoup4==4.13.3
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic>=2.7.4

